# docs/api/openapi-core.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Secure Incident API",
    "version": "0.1.1",
    "description": "Core API draft for the first implementation stage of Secure Incydent. This version follows the domain rule that a user can belong to many projects, but only within one organization or in private projects."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user account",
        "description": "Creates a new user account. After registration the user may create one organization, join one existing organization, or continue with private projects only.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedIdResponse"
                }
              }
            }
          },
          "409": {
            "description": "Username or email already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log in user and create a session",
        "description": "Authenticates the user and establishes a session, for example by setting an HTTP-only cookie.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log out current user",
        "description": "Invalidates the current session.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logged out successfully"
          }
        }
      }
    },
    "/auth/request-password-reset": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request password reset",
        "description": "Initiates password reset flow. The response should not reveal whether the account exists.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Reset request accepted"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password using reset token",
        "description": "Finalizes password reset using a previously issued reset token/code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Invalid or expired reset token"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current authenticated user",
        "description": "Returns the currently authenticated user resolved from the session, including the current organization id if the user belongs to one.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
...
    "/projects/{project_id}/invites": {
      "post": {
        "tags": [
          "Invitations"
        ],
        "summary": "Create project invite",
        "description": "Generates an invite token that lets a user join a specific project directly.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectInviteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project invite created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            }
          }
        }
      }
    },
    "/projects/join": {
      "post": {
        "tags": [
          "Invitations"
        ],
        "summary": "Join project using invite token",
        "description": "Accepts a project invite and joins the current user to the project.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinByInviteRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Joined project successfully"
          }
        }
      }
    },
    "/invites/{token}": {
      "get": {
        "tags": [
          "Invitations"
        ],
        "summary": "Inspect invite before accepting",
        "description": "Lets frontend validate an invite link and preview whether it grants access to an organization or a project.",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite details returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreviewResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invite not found"
          }
        }
      },
      "delete": {
        "tags": ["Invitations"],
        "summary": "Revoke invite",
        "operationId": "revoke_invite",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "token", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Invite revoked" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Not authorized to revoke this invite" },
          "404": { "description": "Invite not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "ProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "username",
          "email",
          "password"
        ],
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/CurrentUserResponse"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "remember_user": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "PasswordResetRequest": {
        "type": "object",
        "required": [
          "email_or_username"
        ],
        "properties": {
          "email_or_username": {
            "type": "string"
          }
        }
      },
      "PasswordResetConfirmRequest": {
        "type": "object",
        "required": [
          "reset_token",
          "new_password"
        ],
        "properties": {
          "reset_token": {
            "type": "string"
          },
          "new_password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "CurrentUserResponse": {
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
      "CreateProjectInviteRequest": {
        "type": "object",
        "required": ["role_id"],
        "properties": {
          "role_id": { "type": "string", "format": "uuid" },
          "expires_at": { "type": "string", "format": "date-time", "nullable": true },
          "max_uses": { "type": "integer", "minimum": 1, "nullable": true }
        }
      },
      "InviteResponse": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "invite_url": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "JoinByInviteRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
      "CreatedIdResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "InvitePreviewResponse": {
        "type": "object",
        "required": [
          "scope",
          "target_name",
          "is_valid"
        ],
        "properties": {
          "scope": {
            "type": "string",
            "enum": [
              "ORGANIZATION",
              "PROJECT"
            ]
          },
          "target_name": {
            "type": "string"
          },
          "is_valid": {
            "type": "boolean"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  }
}